requirements.txt

fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
# Dolphin Tools Dependencies
numpy==1.24.4
omegaconf==2.3.0
opencv-python-headless==4.5.5.64 
pillow==9.3.0
timm==0.5.4
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu # Use CPU version for lighter image, change if GPU needed
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.47.0
accelerate==1.6.0
pymupdf==1.26
